Server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Pillow==10.4.0
numpy==2.1.1
opencv-python==4.10.0.84
scikit-learn==1.5.2